refactor: extract card renderer and drop indicator from large components
