+++ src/utils/terminal-panel-helpers.js
@@ -30,3 +30,61 @@ export class SplitNode {
     this.element = null;
   }
 }
+
+// ===== Pure DOM helpers =====
+
+import { _el } from './dom.js';
+import { computeResizeRatio } from './split-helpers.js';
+import { getPanels } from './split-layout-ops.js';
+
+/** Check if an element is a split container with the given direction. */
+export function isSameDirectionSplit(el, direction) {
+  return el?.classList.contains('split-container') && el.classList.contains(`split-${direction}`);
+}
+
+/** Create a new split container element. */
+export function createSplitContainer(direction, flex = '1') {
+  const el = _el('div', `split-container split-${direction}`);
+  el.style.flex = String(flex);
+  return el;
+}
+
+/** Find the adjacent non-handle panels around a split handle. */
+export function adjacentPanels(handle, splitEl) {
+  const children = Array.from(splitEl.children);
+  const idx = children.indexOf(handle);
+  let before = null, after = null;
+  for (let i = idx - 1; i >= 0; i--) {
+    if (!children[i].classList.contains('split-handle')) { before = children[i]; break; }
+  }
+  for (let i = idx + 1; i < children.length; i++) {
+    if (!children[i].classList.contains('split-handle')) { after = children[i]; break; }
+  }
+  return [before, after];
+}
+
+/** Equalize flex values for all direct panel children of a split container. */
+export function equalizeChildren(splitEl) {
+  for (const panel of getPanels(splitEl)) {
+    panel.style.flex = '1';
+  }
+}
+
+/** Perform resize calculation on two adjacent panels. */
+export function doResize(e, handle, splitEl, direction, fitAllFn) {
+  const [panelBefore, panelAfter] = adjacentPanels(handle, splitEl);
+  if (!panelBefore || !panelAfter) return;
+
+  const { size, start, mouse } = DIRECTION_PROPS[direction];
+  const rectBefore = panelBefore.getBoundingClientRect();
+  const rectAfter = panelAfter.getBoundingClientRect();
+
+  const totalSize = rectBefore[size] + rectAfter[size];
+  const ratio = computeResizeRatio(e[mouse], rectBefore[start], totalSize);
+
+  const totalFlex = (parseFloat(panelBefore.style.flex) || 1) + (parseFloat(panelAfter.style.flex) || 1);
+  panelBefore.style.flex = `${totalFlex * ratio}`;
+  panelAfter.style.flex = `${totalFlex * (1 - ratio)}`;
+
+  fitAllFn();
+}
